Move components to related folders to maintain file structure
