+++ src/Pages/Home.jsx
@@ -1,4 +1,4 @@
-import HomeCarousel from "../HomeCarousel"
+import HomeCarousel from "../Components/Carousels/HomeCarousel"
 import image1 from "../assest/image1.png"
 import image2 from "../assest/image2.png"
 import image3 from "../assest/image3.png"
